Implement rudimentary image loading
This commit introduces the concept of Media and Image nodes to the
Graph. It allows retrieving the original URL from a Media entity or an
entity containing a file field. This can be used to load the profile
picture for a user.

An important part that is omitted from this new functionality is a way
to query the platform for media as well as the ability for the front-end
to load different sizes for an image. Drupal's image styles don't seem
to provide a great fit for an arbitrary front-end application so more
work will be needed.